_processo_Inteligencia/_sistemas_estruturais/genius_harness/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
supabase==2.3.4
langgraph==0.0.26
langchain-openai==0.0.8
python-dotenv==1.0.1
pydantic==2.5.2
